backend/app/usecase/validator/alias.go
package validator

import "regexp"

const (
	customAliasMaxLength = 50
)

// CustomAlias represents format validator for custom alias
type CustomAlias struct {
	uriPattern *regexp.Regexp
}

// IsValid checks whether the given alias has valid format.
func (c CustomAlias) IsValid(alias string) bool {
	if alias == "" {
		return true
	}

	if len(alias) >= customAliasMaxLength {
		return false
	}
	return true
}

// NewCustomAlias creates custom alias validator.
func NewCustomAlias() CustomAlias {
	return CustomAlias{}
}
